% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\alias{anova_oneway}
\title{ANOVA}
\usage{
anova_oneway(x, groups)
}
\arguments{
\item{x}{numeric matrix.}

\item{groups}{- a vector giving groups for each column of x.}
}
\value{
a data.frame where each row contains the results of an anova test
performed on the corresponding row of x.
}
\description{
Performs an analysis of variance test on each row of the input matrix.
}
\details{
Functions to perform ANOVA analysis for rows of matrices.

\code{anova_oneway} - one-way anova. Same as \code{aov(x ~ groups)}
}
\examples{
anova_oneway(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{aov()}, \code{anova()}
}
\author{
Karolis Koncevičius
}
