% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{linearmodel}
\alias{linearmodel}
\alias{row_lm_f}
\alias{col_lm_f}
\title{F-test for nested linear regression models}
\usage{
row_lm_f(
  x,
  m,
  null = stats::model.matrix(~1, data = data.frame(seq_len(nrow(m))))
)

col_lm_f(
  x,
  m,
  null = stats::model.matrix(~1, data = data.frame(seq_len(nrow(m))))
)
}
\arguments{
\item{x}{numeric matrix.}

\item{m}{- a model matrix for a linear regression model to be tested.}

\item{null}{- a null model which the original model will be compared against (default = intercept-only).}
}
\value{
a data.frame where each row contains the results of an F-test
performed on the corresponding row/column of x.\cr\cr
Each row contains the following information (in order):\cr
1. obs - total number of observations\cr
2. rsquared.model - R-squared of the full model\cr
3. rsquared.null - R-squared of the null model\cr
4. df.model - model terms degrees of freedom\cr
5. df.residual - residual degrees of freedom\cr
6. statistic - F statistic\cr
7. pvalue - p-value\cr
}
\description{
Performs an F-test comparing two nested linear regression models on each row/column of the input matrix.
}
\details{
\code{row_lm_f} - F-test for linear regression on rows.
\code{col_lm_f} - F-test for linear regression on columns.
}
\examples{
X   <- t(iris[,-5])
mod <- model.matrix(~ iris$Species)
row_lm_f(X, mod)

# with specified null model
X <- mtcars[,c("mpg", "disp", "qsec")]
mod  <- model.matrix(~ mtcars$drat + mtcars$cyl + mtcars$wt + mtcars$vs)
mod0 <- model.matrix(~               mtcars$cyl + mtcars$wt + mtcars$vs)
col_lm_f(X, mod, mod0)

}
\seealso{
\code{lm()}, \code{anova()}
}
\author{
Karolis KonceviÄius
}
