% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\alias{row.oneway.equalvar}
\alias{row.oneway.welch}
\title{ONEWAY ANOVA}
\usage{
row.oneway.equalvar(x, groups)

row.oneway.welch(x, groups)
}
\arguments{
\item{x}{numeric matrix.}

\item{groups}{a vector specifying group membership for each column of x.}
}
\value{
a data.frame where each row contains the results of an oneway anova
test performed on the corresponding row of x.
}
\description{
Performs an analysis of variance tests on each row of the input matrix.
}
\details{
Functions to perform ONEWAY ANOVA analysis for rows of matrices.

\code{row.oneway.equalvar} - one-way anova. Same as \code{aov(x ~ groups)}
\code{row.oneway.welch} _ one-way anova with Welch correction for variances.
Same as \code{oneway.test(var.equal=FALSE)}
}
\examples{
row.oneway.equalvar(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{aov()}, \code{oneway.test()}
}
\author{
Karolis Koncevičius
}
