% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal.R
\name{oneway_equalvar}
\alias{oneway_equalvar}
\title{Kruskal-Wallis Rank Sum Test}
\usage{
oneway_equalvar(x, groups)
}
\arguments{
\item{x}{numeric matrix.}

\item{groups}{a vector specifying group membership for each column of x.}
}
\value{
a data.frame where each row contains the results of an oneway anova
test performed on the corresponding row of x.
}
\description{
Performs a Kruskal-Wallis rank sum test on each row of the input matrix.
}
\details{
\code{kruskal} - sam as as \code{kruskal.test(0}
}
\examples{
oneway_equalvar(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{aov()}, \code{oneway.test()}
}
\author{
Karolis Koncevičius
}
