% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlet.R
\name{bartlett}
\alias{bartlett}
\title{Bartlett test}
\usage{
bartlett(x, groups)
}
\arguments{
\item{x}{numeric matrix.}

\item{groups}{a vector specifying group membership for each column of x.}
}
\value{
a data.frame where each row contains the results of the bartlett test
performed on the corresponding row of x.
}
\description{
Performs the Bartlett's test of homogeneity of variances on each row of the
input matrix.
}
\details{
Function to perform Bartlett's tests for homogeneity of variances in each of
the groups on the rows of a matrix.

NA values are always ommited. If values are missing for a whole group - that
group is discarded. Groups with only one observation are also discarded.

\code{bartlett(x, groups)} - Bartlet's test.
Same as \code{bartlett.test(x,  groups)}
}
\examples{
bartlett(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{bartlett.test()}
}
\author{
Karolis Koncevičius
}
