% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal.R
\name{kruskalwallis}
\alias{kruskalwallis}
\title{Kruskal-Wallis Rank Sum Test}
\usage{
kruskalwallis(x, groups)
}
\arguments{
\item{x}{numeric matrix.}

\item{groups}{a vector specifying group membership for each column of x.}
}
\value{
a data.frame where each row contains the results of a Kruskal-Wallis
test performed on the corresponding row of x.
}
\description{
Performs a Kruskal-Wallis rank sum test on each row of the input matrix.
}
\details{
\code{kruskalwallis} - sam as as \code{kruskal.test()}
}
\examples{
kruskalwallis(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{kruskal.test()}
}
\author{
Karolis Koncevičius
}
