% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\alias{row.oneway.equalvar}
\alias{col.oneway.equalvar}
\alias{row.oneway.welch}
\alias{col.oneway.welch}
\title{ONEWAY ANOVA}
\usage{
row.oneway.equalvar(x, g)

col.oneway.equalvar(x, g)

row.oneway.welch(x, g)

col.oneway.welch(x, g)
}
\arguments{
\item{x}{numeric matrix.}

\item{g}{a vector specifying group membership for each observation of x.}
}
\value{
a data.frame where each row contains the results of an oneway anova
test performed on the corresponding row/column of x.
The columns will vary depending on the type of test performed.\cr\cr
They will contain a subset of the following information:\cr
1. obs.tot - total number of observations\cr
2. obs.group - number of groups\cr
3. sum.sq.between - between group (treatment) sum of squares\cr
4. sum.sq.within - within group (residual) sum of squares\cr
5. mean.sq.between - between group mean squares\cr
6. mean.sq.within - within group mean squares\cr
7. df.between - between group (treatment) degrees of freedom\cr
8. df.within - within group (residual) degrees of freedom\cr
9. statistic.F - F statistic\cr
10. p.value - p.value
}
\description{
Performs an analysis of variance tests on each row/column of the input matrix.
}
\details{
Functions to perform ONEWAY ANOVA analysis for rows/columns of matrices.

\code{row.oneway.equalvar}, \code{col.oneway.equalvar}
- one-way anova on columns. Same as \code{aov(x ~ group)}

\code{row.oneway.welch}, \code{col.oneway.welch}
- one-way anova with Welch correction on columns. Same as \code{oneway.test(var.equal=FALSE)}
}
\examples{
col.oneway.welch(iris[,1:4], iris$Species)
row.oneway.equalvar(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{aov()}, \code{oneway.test()}
}
\author{
Karolis Koncevičius
}
