% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\alias{oneway_equalvar}
\alias{oneway_welch}
\title{ONEWAY ANOVA}
\usage{
oneway_equalvar(x, groups)

oneway_welch(x, groups)
}
\arguments{
\item{x}{numeric matrix.}

\item{groups}{a vector specifying group membership for each column of x.}
}
\value{
a data.frame where each row contains the results of an oneway anova
test performed on the corresponding row of x.
}
\description{
Performs an analysis of variance tests on each row of the input matrix.
}
\details{
Functions to perform ONEWAY ANOVA analysis for rows of matrices.

\code{oneway_equalvar} - one-way anova. Same as \code{aov(x ~ groups)}
\code{oneway_welch} _ one-way anova with Welch correction for variances.
Same as \code{oneway.test(var.equal=FALSE)}
}
\examples{
oneway_equalvar(t(iris[,1:4]), iris$Species)

}
\seealso{
\code{aov()}, \code{oneway.test()}
}
\author{
Karolis Koncevičius
}
